Fixed seg fault
Was passing the double pointer by reference, removing & fixed that as I should just pass it by value given that it's a memory address. Size has also been reduced from 1,000,000 to 1024.

Compiles and runs, but now every value except A1,1 is -nan, so something is still wrong but it's much closer.