Almost working
For some reason in calculateQ in lanczos.cpp 'i' is initialized as 21845 rather than 1 as I specified. Not sure how to fix this issue, but because of this the array is accessing the wrong memory so it is uninitialized values that it runs my algorithm with, so my resulting values are garbage.